Change file encoding to UTF-8 to ensure Xtend compatibility
I have had to change the encoding setting to make it explicit that the
files should be considered UTF-8. This is because there is now a fair
amount of Xtend code, some of which uses template strings with
guillemets in them. If I don’t make this change, I get lots of weird
compilation issues.

Signed-off-by: Steffen Zschaler <steffen.zschaler@kcl.ac.uk>